refactor: Extract appointment form logic into a custom hook for improved modularity and maintainability
