+++ src/app/(core)/appointments/components/add-appointment-button.tsx
@@ -31,7 +31,6 @@ const AddAppointmentButton = ({
       <AddAppointmentForm
         patients={patients}
         doctors={doctors}
-        isOpen={isOpen}
         onSuccess={() => setIsOpen(false)}
       />
     </Dialog>
